New second ship
Changed the secondShip so that it appears flipped beneath you
Turned off trail particles when second ship is activated
Made it so the second ship could be killed by bullet
Made it so the second ship could be killed by enemy
Made it so the second ship could activate slowtime
Made it so the second ship could pick up shield emblems
Made it so the second ship could pick up power emblems
The collisions for the second ship are slightly off, not sure why and
will fix when I get the chance.
Haven't edited the camera bounds, will fix when I get the chance.

BTW I added a new gameobject in the PlayerShipController to pick up the
second ship, sorry if there was already one before. I'll check when i
get the chance.